refactor(agentPresenter): move utility module
