+++ src/main/presenter/agentPresenter/utility/index.ts
@@ -1,1 +1,2 @@
-export {}
+export { UtilityHandler } from './utilityHandler'
+export { enhanceSystemPromptWithDateTime } from './promptEnhancer'
