+++ src/main/presenter/sessionPresenter/index.ts
@@ -29,7 +29,7 @@ import { LLMEventHandler } from '../agentPresenter/streaming/llmEventHandler'
 import { SearchHandler } from '../searchPresenter/handlers/searchHandler'
 import { StreamGenerationHandler } from '../agentPresenter/streaming/streamGenerationHandler'
 import { PermissionHandler } from '../agentPresenter/permission/permissionHandler'
-import { UtilityHandler } from './utility/utilityHandler'
+import { UtilityHandler } from '../agentPresenter/utility/utilityHandler'
 import { CommandPermissionService } from '../permission/commandPermissionService'
 import type { ThreadHandlerContext } from '../searchPresenter/handlers/baseHandler'
 import { ConversationManager, type CreateConversationOptions } from './managers/conversationManager'
